backend/requirements.txt
flask
flask-cors
python-dotenv

# ASR + audio
openai-whisper
torch
ffmpeg-python

# Summarization / NLP
transformers
sentencepiece
accelerate
